up to (and including) 4.23

+++ part4/controllers/blogs.js
@@ -1,6 +1,9 @@
+const jwt = require('jsonwebtoken')
+
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const { userExtractor } = require('../utils/middleware')
 
 blogRouter.get('/', async (request, response) =>{
     const blogs = await Blog.find({})
@@ -17,9 +20,16 @@ blogRouter.get('/:id', async (request, response) =>{
     }
 })
 
-blogRouter.post('/', async (request, response) => {
-    const blog = new Blog(request.body)
-    const user = await User.findById("649d162d7b838dd2c4f5f43a")
+blogRouter.post('/', userExtractor, async (request, response) => {
+    const body = request.body
+    const blog = new Blog({
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        likes: body.likes,
+        user: request.user.id
+    })
+    const user = request.user
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
@@ -27,7 +37,12 @@ blogRouter.post('/', async (request, response) => {
     response.status(201).json(savedBlog)
 })
 
-blogRouter.delete('/:id', async (request, response) => {
+blogRouter.delete('/:id', userExtractor, async (request, response) => {
+    const blog = await Blog.findById(request.params.id)
+    if ( blog && blog.user.toString() !== request.user.id ) {
+        return response.status(403).json({ error: "user doesn't own this note"})
+    }
+
     const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
 })
